Rename `BasicMapping` to `TypeMappingImpl`
