+++ README.md
@@ -52,14 +52,14 @@ If there is no predefined mapping for the database, mapping for MySQL 5.7 will b
 
 If the automatically selected mapping doesn't match your requirements, a custom mapping can be
 specified during creation of `JdbcStorageFactory`. There is the basic builder for this purpose - 
-`BasicMapping.basicBuilder()`. This builder contains mappings for all types
+`TypeMappingImpl.basicBuilder()`. This builder contains mappings for all types
 (equal to MySQL 5.7 mapping). So only required types should be overridden:
 
 ```
-TypeMapping mapping = BasicMapping.basicBuilder()
-                                  .add(Type.INT, "INT4")
-                                  .add(Type.LONG, "INT8")
-                                  .build();
+TypeMapping mapping = TypeMappingImpl.basicBuilder()
+                                     .add(Type.INT, "INT4")
+                                     .add(Type.LONG, "INT8")
+                                     .build();
 JdbcStorageFactory.newBuilder()
                   .setTypeMapping(mapping)
 ```
