+++ rdbms/src/main/java/io/spine/server/storage/jdbc/JdbcStorageFactory.java
@@ -228,7 +228,7 @@ public class JdbcStorageFactory implements StorageFactory {
         /**
          * Sets {@link TypeMapping}, which defines {@link Type} names for the database used.
          *
-         * <p>Use the {@linkplain BasicMapping#basicBuilder() basic builder}
+         * <p>Use the {@linkplain TypeMappingImpl#basicBuilder() basic builder}
          * to build a custom mapping.
          *
          * <p>If the mapping was not specified, it is
