+++ rdbms/src/main/java/io/spine/server/storage/jdbc/PredefinedMapping.java
@@ -26,8 +26,8 @@ import io.spine.type.TypeName;
 import java.sql.DatabaseMetaData;
 import java.sql.SQLException;
 
-import static io.spine.server.storage.jdbc.BasicMapping.basicBuilder;
 import static io.spine.server.storage.jdbc.Type.BYTE_ARRAY;
+import static io.spine.server.storage.jdbc.TypeMappingImpl.basicBuilder;
 
 /**
  * Predefined {@linkplain TypeMapping type mappings} for different databases.
@@ -46,7 +46,7 @@ public enum PredefinedMapping implements TypeMapping {
     private final int minorVersion;
 
     PredefinedMapping(String databaseProductName, int majorVersion,
-                      int minorVersion, BasicMapping.Builder mappingBuilder) {
+                      int minorVersion, TypeMappingImpl.Builder mappingBuilder) {
         this.databaseProductName = databaseProductName;
         this.majorVersion = majorVersion;
         this.minorVersion = minorVersion;
